backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
paddlepaddle
paddleocr
numpy
opencv-python-headless
Pillow
